+++ src/main/resources/db/migration/V3__adicao_tabela_provento.sql
@@ -1,7 +1,7 @@
 create table provento(
 	id_provento int auto_increment not null,
 	data_provento date not null,
-	valor float not null,
+	valor decimal(14,2) not null,
 	id_ativo int not null,
 	constraint pk_provento primary key(id_provento),
 	constraint fk_provento_ativo foreign key (id_ativo) references ativo(id_ativo)
